+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Unreleased
 
 -->
 
+## 1.0.1 - 2026-06-26
+- Fix unsuccessful fetch of the capabilities document from WMS server
+- Only apply the result limit after filtering results by perimeter
+
 ## 1.0.0 - 2026-04-29
 - Fix missing entries or non-sorted results when searching for locations
 - Reduce and simplify GitHub workflow triggers, add some clarifications